feat: Update block property in ITransactionLog interface to include timestamp
Update the `block` property in the `ITransactionLog` interface to include a `timestamp` field. This change is necessary to provide additional information about the block associated with a transaction. The `timestamp` field will be of type `bigint` to accurately represent the timestamp value.

Refactor the `transaction.model.ts` file in the `transaction-search` project to include the updated interface definition.