Angular 17 upgrade (#480)
* First version of migration

* Prettier updated

* feat: Update dependencies and fix imports in bid-store, document-locker-store, payment-request-store, transaction-store, and vaults-store effects

* refactor: Update IReceiptTransaction interface to use bigint for status and type properties

* Migration issues fixing

* feat: Update block property in ITransactionLog interface to include timestamp

Update the `block` property in the `ITransactionLog` interface to include a `timestamp` field. This change is necessary to provide additional information about the block associated with a transaction. The `timestamp` field will be of type `bigint` to accurately represent the timestamp value.

Refactor the `transaction.model.ts` file in the `transaction-search` project to include the updated interface definition.

* chore: Update getCurrentPrice function to use async/await syntax

Refactor the `getCurrentPrice` function in the `PriceFeedService` class to use the async/await syntax instead of Promise chaining. This change improves the readability and maintainability of the code by eliminating nested callbacks and reducing the number of lines.

* refactor: Update `TokensService` to use `Contract` type from `web3`

* refactor: Update `transferNativeTokenSC` method in `TokensService` to return converted receipt

Refactor the `transferNativeTokenSC` method in the `TokensService` class to return a converted receipt object of type `IReceiptTransaction`. This change ensures that the receipt object has consistent data types for properties like `blockNumber`, `transactionIndex`, `cumulativeGasUsed`, `effectiveGasPrice`, `gasUsed`, and `type`. By converting the receipt, we can provide more accurate and reliable information about the transaction.

Co-authored-by: dependabot[bot] <support@github.com>

* chore: Update web3 dependency to version 4.10.0

Update the web3 dependency in the package.json files of the token-bridge, transaction-search, and web3-login projects to version 4.10.0. This update ensures compatibility with the latest version of web3 and allows for improved functionality and bug fixes.

* chore: Update web3 dependency to version 4.10.0

* chore: Update dependencies to version 17.0.1

* chore: Update package.json version to 0.0.4

* chore: Update node version to 18.x in docker-deploy-dev.yml

* chore: Update node version to 18.x in docker-deploy-prod.yml, docker-deploy-test.yml, local-libraries.yml, and verifications.yml

Update the node version to 18.x in the docker-deploy-prod.yml, docker-deploy-test.yml, local-libraries.yml, and verifications.yml files. This update ensures compatibility with the latest version of Node.js and allows for improved functionality and bug fixes.

* chore: Update node version to 18.x in verifications.yml

Update the node version to 18.x in the verifications.yml file. This update ensures compatibility with the latest version of Node.js and allows for improved functionality and bug fixes.

* refactor: Update `BidService` to use web3 provider from `web3ProviderService`

Refactor the `BidService` class to use the web3 provider from the `web3ProviderService` instead of creating a new instance of `Web3` with a hardcoded URL. This change ensures consistency and improves maintainability by centralizing the web3 provider configuration.