```text feat: Add support for deploying Apollo module with communityVaultAddress
This commit updates the Apollo module to accept the communityVaultAddress as a parameter when building the module. It also updates the deploy script to pass the communityVaultAddress to the Apollo module during deployment. This change enhances the functionality of the application by allowing the Apollo module to interact with the community vault contract.