feat: Add support for adding tokens to wallet
This commit adds support for adding tokens to the wallet in the token management page. It introduces the 'addTokenToWallet' action and updates the TokenSetupService to handle the 'wallet_watchAsset' method for adding ERC20 tokens to the wallet. This change enhances the functionality of the application by allowing users to easily add tokens to their wallet.